refactor the fat guy
