+++ ca-grocery-receipt-Selenium/Program.cs
@@ -20,7 +20,7 @@ class Program
 
         ProjectDTO projectConfig = new ProjectDTO()
         {
-            receiptCode = "122623 165625 1698 02809"
+            receiptCode = "122623 155625 1698 02809"
         };
         //////// end test case setting
