+++ ca-grocery-receipt-Selenium/Utilities/JobList.cs
@@ -8,6 +8,8 @@ public static class JobList
         List<BaseWorkItem> todayJobList = new List<BaseWorkItem>() { };
         todayJobList.Add(new WebpageWorker.FillTheReceiptCode() { jobSequence = 10 });
         todayJobList.Add(new WebpageWorker.FillTheForm() { jobSequence = 20 });
+        todayJobList.Add(new WebpageWorker.FillTheContactInformation() { jobSequence = 30 });
+        todayJobList.Add(new WebpageWorker.SubmitTheForm() { jobSequence = 40 });
         return todayJobList.OrderBy(seq => seq.jobSequence).ToList();
     }
 }
